+++ src/test/java/io/jtest/utils/matcher/FlowMatcherTests.java
@@ -6,7 +6,13 @@ import com.fasterxml.jackson.databind.node.TextNode;
 import io.jtest.utils.matcher.condition.MatchCondition;
 import org.junit.jupiter.api.Test;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;